Export sprites and styles
