--- sprites/dark-v8/sprite.json
@@ -1,16 +1,0 @@
-{
-  "default_marker": {
-    "width": 20,
-    "height": 50,
-    "x": 0,
-    "y": 0,
-    "pixelRatio": 1
-  },
-  "secondary_marker": {
-    "width": 20,
-    "height": 50,
-    "x": 20,
-    "y": 0,
-    "pixelRatio": 1
-  }
-}
