--- sprites/dark-v8/sprite@2x.json
@@ -1,16 +1,0 @@
-{
-  "default_marker": {
-    "width": 40,
-    "height": 100,
-    "x": 0,
-    "y": 0,
-    "pixelRatio": 2
-  },
-  "secondary_marker": {
-    "width": 40,
-    "height": 100,
-    "x": 40,
-    "y": 0,
-    "pixelRatio": 2
-  }
-}
